chore(types): moving types to the relevent packages and deleting types.go
